Sun, Jun 2, 2024, 4:10 PM -03:00

+++ py/week1/problems/interpreter.py
@@ -1,7 +1,7 @@
 x, y, z = input("Expression: ").split(" ")
 
-x = int(x)
-z = int(z)
+x = float(x)
+z = float(z)
 
 match y:
     case "+":
